+++ SLIP_LU/Demo/example4.c
@@ -18,8 +18,8 @@
 // example4 > out
 // out is file for output calculated result
 
-int32_t Ap[5] = {0, 3, 5, 8, 11};
-int32_t Ai[11]       = {0, 1, 2, 2, 3, 1, 2, 3, 0, 1,  2};
+int64_t Ap[5] = {0, 3, 5, 8, 11};
+int64_t Ai[11]       = {0, 1, 2, 2, 3, 1, 2, 3, 0, 1,  2};
 double Axnum[11] = {1, 2, 7, 1, 2, 4, 1, 3, 1, 12, 1};    // Numerator of x
 double Axden[11] = {3, 3, 6, 1, 7, 1, 1, 1, 5, 1,  1};    // Denominator of x
 double bxnum[4] = {17, 182, 61, 67};                      // Numerator of b
@@ -54,11 +54,11 @@ int main (void)
     //--------------------------------------------------------------------------
 
     SLIP_info ok;
-    int32_t n = 4, nz = 11, numRHS = 2;
+    int64_t n = 4, nz = 11, numRHS = 2;
     double* x = NULL;
     double **b_doub = NULL;
-    int32_t* i = NULL;
-    int32_t* p = NULL;
+    int64_t* i = NULL;
+    int64_t* p = NULL;
     double** soln = NULL;
     SLIP_sparse* A = NULL ;
     SLIP_dense *b = NULL ;
@@ -71,8 +71,8 @@ int main (void)
 
     x = (double*) SLIP_malloc(nz* sizeof(double));
     b_doub = SLIP_create_double_mat(n, numRHS);
-    i = (int32_t*) SLIP_malloc(nz* sizeof(int32_t));
-    p = (int32_t*) SLIP_malloc((n+1)* sizeof(int32_t));
+    i = (int64_t*) SLIP_malloc(nz* sizeof(int64_t));
+    p = (int64_t*) SLIP_malloc((n+1)* sizeof(int64_t));
     soln = SLIP_create_double_mat(n, numRHS);;
     if (!x || !b_doub || !i || !p || !soln || !option)
     {
@@ -80,14 +80,14 @@ int main (void)
         FREE_WORKSPACE;
         return 0;
     }
-    for (int32_t j = 0; j < n; j++)                           // Get b & p
+    for (int64_t j = 0; j < n; j++)                           // Get b & p
     {
         p[j] = Ap[j];
         b_doub[j][0] = bxnum[j]/bxden[j];
         b_doub[j][1] = bxnum2[j]/bxden2[j];
     }
     p[n] = Ap[n];
-    for (int32_t j = 0; j < nz; j++)                          // Get i and x
+    for (int64_t j = 0; j < nz; j++)                          // Get i and x
     {
         i[j] = Ai[j];
         x[j] = Axnum[j]/Axden[j];
