+++ SLIP_LU/Demo/example5.c
@@ -55,7 +55,7 @@ int main (int argc, char **argv)
     //--------------------------------------------------------------------------
 
     SLIP_info ok;
-    int32_t n = 177, numRHS = 500;
+    int64_t n = 177, numRHS = 500;
     double **b_doub = NULL;
     double **soln = NULL;
     SLIP_sparse *A = NULL ;
@@ -81,11 +81,11 @@ int main (int argc, char **argv)
         FREE_WORKSPACE;
         return 0;
     }
-    int32_t seed = 12;      // Just an arbitrary random seed
+    unsigned int seed = 12;      // Just an arbitrary random seed
     srand (seed);
-    for (int32_t i = 0; i < n; i++)
+    for (int64_t i = 0; i < n; i++)
     {
-        for (int32_t j = 0; j < n; j++)
+        for (int64_t j = 0; j < n; j++)
         {
             b_doub[i][j] = rand();
         }
