+++ SLIP_LU/MATLAB/Source/slip_mex_output_col_permut.c
@@ -8,14 +8,14 @@
 
 //------------------------------------------------------------------------------
 
-/* Purpose: This function outputs an int32_t array as a mxArray.*/
+/* Purpose: This function outputs an int64_t array as a mxArray.*/
 
 #include "SLIP_LU_mex.h"
 
 mxArray* slip_mex_output_col_permut
 (
-    int32_t* x,        // int32_t array to be output
-    int32_t n          // size of x
+    int64_t* x,        // int64_t array to be output
+    int64_t n          // size of x
 )
 {
     // Create a n*1 matlab array
@@ -25,7 +25,7 @@ mxArray* slip_mex_output_col_permut
     double* x2 = mxGetPr(Xmatlab);
 
     // Get cast x as double
-    for (int32_t k = 0; k < n; k++)
+    for (int64_t k = 0; k < n; k++)
     {
         x2[k] = (double) x[k];
     }
