+++ SLIP_LU/Source/SLIP_build_sparse_trip_mpz.c
@@ -20,11 +20,11 @@
 SLIP_info SLIP_build_sparse_trip_mpz
 (
     SLIP_sparse **A_handle,     // matrix to construct
-    int32_t *I,         // set of row indices
-    int32_t *J,         // set of column indices
+    int64_t *I,         // set of row indices
+    int64_t *J,         // set of column indices
     mpz_t *x,           // Set of values in full precision integer
-    int32_t n,          // dimension of the matrix
-    int32_t nz          // number of nonzeros in A (size of x, I, and J vectors)
+    int64_t n,          // dimension of the matrix
+    int64_t nz          // number of nonzeros in A (size of x, I, and J vectors)
 )
 {
