+++ SLIP_LU/Source/SLIP_create_mpfr_mat.c
@@ -16,8 +16,8 @@
 
 mpfr_t** SLIP_create_mpfr_mat
 (
-    int32_t m,     // number of rows
-    int32_t n,     // number of columns
+    int64_t m,     // number of rows
+    int64_t n,     // number of columns
     SLIP_options *option// command options containing the prec for mpfr
 )
 {
@@ -35,11 +35,11 @@ mpfr_t** SLIP_create_mpfr_mat
 
     mpfr_t **x = (mpfr_t**) SLIP_calloc(m, sizeof(mpfr_t*));
     if (!x) {return NULL;}
-    for (int32_t i = 0; i < m; i++)
+    for (int64_t i = 0; i < m; i++)
     {
         x[i] = NULL;
     }
-    for (int32_t i = 0; i < m; i++)
+    for (int64_t i = 0; i < m; i++)
     {
         x[i] = (mpfr_t*) SLIP_calloc(n, sizeof(mpfr_t));
         if (x[i] == NULL)
@@ -48,7 +48,7 @@ mpfr_t** SLIP_create_mpfr_mat
             SLIP_delete_mpfr_mat(&x, m, n);
             return NULL;
         }
-        for (int32_t j = 0; j < n; j++)
+        for (int64_t j = 0; j < n; j++)
         {
             if (SLIP_mpfr_init2(x[i][j], option->prec) != SLIP_OK)
             {
