+++ SLIP_LU/Source/SLIP_create_mpq_mat.c
@@ -14,8 +14,8 @@
 
 mpq_t** SLIP_create_mpq_mat
 (
-    int32_t m,     // number of rows
-    int32_t n      // number of columns
+    int64_t m,     // number of rows
+    int64_t n      // number of columns
 )
 {
 
@@ -33,7 +33,7 @@ mpq_t** SLIP_create_mpq_mat
     // Malloc space
     mpq_t **x = (mpq_t**) SLIP_calloc(m, sizeof(mpq_t*));
     if (!x) {return NULL;}
-    for (int32_t i = 0; i < m; i++)
+    for (int64_t i = 0; i < m; i++)
     {
         x[i] = (mpq_t*) SLIP_calloc(n, sizeof(mpq_t));
         if ( x[i] == NULL)
@@ -42,7 +42,7 @@ mpq_t** SLIP_create_mpq_mat
             SLIP_delete_mpq_mat(&x, m, n);
             return NULL;
         }
-        for (int32_t j = 0; j < n; j++)
+        for (int64_t j = 0; j < n; j++)
         {
             if (SLIP_mpq_init(x[i][j]) != SLIP_OK)
             {
