+++ SLIP_LU/Source/slip_cumsum.c
@@ -16,13 +16,13 @@
 
 SLIP_info slip_cumsum
 (
-    int32_t *p,      // vector to store the sum of c
-    int32_t *c,      // vector which is summed
-    int32_t n        // size of c
+    int64_t *p,      // vector to store the sum of c
+    int64_t *c,      // vector which is summed
+    int64_t n        // size of c
 )
 {
 
-    int32_t i, nz = 0 ;
+    int64_t i, nz = 0 ;
     for (i = 0 ; i < n ; i++)
     {
         p [i] = nz ;
